+++ source/menus.lua
@@ -10,7 +10,19 @@ function menus.DrawMainMenu()
 
 	-- try to centre the Slab window
 	-- note: Border is the border between the window and the layout
-	Slab.BeginWindow('MainMenu', {Title = "Main menu " .. gstrGameVersion,X=fltSlabWindowX,Y=fltSlabWindowY,W=intSlabWidth,H=intSlabHeight,Border=0,AutoSizeWindow=false, AllowMove=false,AllowResize=false,NoSavedSettings=true})
+	local mainMenuOptions = {
+		Title = "Main menu " .. gstrGameVersion,
+		X = fltSlabWindowX,
+		Y = fltSlabWindowY,
+		W = intSlabWidth,
+		H = intSlabHeight,
+		Border = 0,
+		AutoSizeWindow=false,
+		AllowMove=false,
+		AllowResize=false,
+		NoSavedSettings=true
+	}
+	Slab.BeginWindow('MainMenu', mainMenuOptions)
 
 	Slab.BeginLayout("MMLayout",{AlignX="center",AlignY="center",AlignRowY="center",ExpandW=false,Columns = 2})
 		
@@ -76,19 +88,32 @@ function menus.DrawMainMenu()
 		
 		if not gbolIsAHost then
 			Slab.Text("Join on port:" )
-			if Slab.Input('HostEndPoint',{ReturnOnText=true,W=100,Text = ConnectedToPort}) then
+			local joinPortOptions = {
+				ReturnOnText=true,
+				W=100,
+				Text=ConnectedToPort,
+				NumbersOnly=true,
+				NoDrag=true,
+				MinNumber=6000,
+				MaxNumber=6999
+			}
+			if Slab.Input('HostEndPoint', joinPortOptions) then
 				ConnectedToPort = Slab.GetInputText()
 			end
 			
 			if Slab.Button("Join game",{W=155}) then
-				gbolIsAHost = false
-				gbolIsAClient = true
+				if ConnectedToPort == nil then
+					-- Invalid Port number
+					lovelyToasts.show("Error: You must set a port number", 1.5, "bottom")
+				else
+					gbolIsAHost = false
+					gbolIsAClient = true
 
-				ss.ConnectToHost(_, ConnectedToPort)
-				
-				ss.AddItemToClientOutgoingQueue(message)
-				fun.AddScreen("World")
-
+					ss.ConnectToHost(_, ConnectedToPort)
+					
+					ss.AddItemToClientOutgoingQueue(message)
+					fun.AddScreen("World")
+				end
 			end
 			Slab.NewLine()		
 		end
@@ -123,7 +148,16 @@ function menus.DrawCredits()
 	local fltSlabWindowX = love.graphics.getWidth() / 2 - intSlabWidth / 2
 	local fltSlabWindowY = love.graphics.getHeight() / 2 - intSlabHeight / 2
 
-	Slab.BeginWindow('creditsbox',{Title ='About',BgColor = {0.5,0.5,0.5},AutoSizeWindow = true,NoOutline=true,AllowMove=false,X=fltSlabWindowX,Y=fltSlabWindowY})
+	local creditBoxOptions = {
+		Title ='About',
+		BgColor = {0.5,0.5,0.5},
+		AutoSizeWindow = true,
+		NoOutline = true,
+		AllowMove = false,
+		X = fltSlabWindowX,
+		Y = fltSlabWindowY
+	}
+	Slab.BeginWindow('creditsbox', creditBoxOptions)
 	Slab.BeginLayout('mylayout', {AlignX = 'center',Columns = 2})
 
 		Slab.Text("Mars Lander")